+++ src/objects/Turtle.h
@@ -197,7 +197,7 @@ public:
         popMatrix(MODEL);
     }
 
-    void update(int deltaTime) override {
+    void update(float deltaTime) override {
         if (animationEnabled) {
             GLint currentTime = glutGet(GLUT_ELAPSED_TIME);
             prevTime = currentTime;
